+++ src/app/api/cron/wiki/scraper/skills/parse.ts
@@ -14,6 +14,9 @@ export function skillDataParse($: CheerioAPI): {
   // Not every skill has a cooldown
   const hasCooldown = wholeDescription.includes('Cooldown:')
 
+  // Hunter skills have the MARK text after the cooldown
+  const hasMark = wholeDescription.includes('MARK:')
+
   if (!hasCooldown) {
     return {
       description: wholeDescription.trim(),
@@ -22,9 +25,16 @@ export function skillDataParse($: CheerioAPI): {
   }
 
   // The description is everything before the two line breaks and the text Cooldown:
-  const description = wholeDescription.split('Cooldown:')[0].trim()
+  let description = wholeDescription.split('Cooldown:')[0].trim()
   // The cooldown is everything starting with the Cooldown: text
-  const cooldown = wholeDescription.split('Cooldown:')[1].trim()
+  let cooldown = wholeDescription.split('Cooldown:')[1].trim()
+
+  if (hasMark) {
+    // Need to remove the MARK text from the cooldown, then append it to the description
+    const mark = cooldown.split('MARK:')[1].trim()
+    cooldown = cooldown.split('MARK:')[0].trim()
+    description += `\n\nMARK: ${mark}`
+  }
 
   return {
     description,